#13 notification configuration tests
